backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
selenium==4.25.0
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
google-generativeai==0.8.3
python-dotenv==1.0.1
pillow==11.0.0
python-multipart==0.0.12
pydantic==2.9.2
